+++ R/estimation_lin.R
@@ -1,11 +1,9 @@
 estimation_lin = function(obj,coefficient){
   coeff.est = c()
-  iterations = obj$obj$iterations
   posterior = obj$coefficients
   for(j in 1:dim(posterior$GS.beta)[2])
   {
     t1=as.matrix(posterior$GS.beta[,j])
-    t1 = t1[seq(iterations/2+1, iterations,1),]
     coeff.est[j] = stats::median(t1)
   }
   MSE = mean((coeff.est-coefficient)^2)