+++ R/pqrBayes.R
@@ -1,27 +1,27 @@
 #' fit Bayesian penalized quantile regression for linear and varying coefficient models
 #' 
 #' @keywords models
-#' @param g the matrix of predictors (subject to selection) without intercept.
+#' @param g the matrix of predictors (subject to selection). Users do not need to specify an intercept which will be automatically included. 
 #' @param y the response variable. The current version only supports the continuous response.
 #' @param u a vector of effect modifying variable of the quantile varying coefficient model. When fitting a linear model, u = NULL.
-#' @param e a matrix of clinical covariates (including the intercept if specified) not subject to selection.
+#' @param e a matrix of clinical covariates not subject to selection.
 #' @param quant the quantile level specified by users. The default value is 0.5.
 #' @param iterations the number of MCMC iterations. The default value is 10,000.
-#' @param kn the number of interior knots for B-spline. When fitting a linear model, kn = NULL.
-#' @param degree the degree of B-spline basis. When fitting a linear model, degree = NULL.
+#' @param burn.in the number of burn-in iterations. If NULL, the first half of MCMC iterations will be used as burn-ins.
+#' @param spline a list of the number of interior knots (kn) for B-spline and the degree of B-spline basis (degree). When fitting a linear model, spline = NULL.
 #' @param robust logical flag. If TRUE, robust methods will be used. Otherwise, non-robust methods will be used. The default value is TRUE.
 #' @param sparse logical flag. If TRUE, spike-and-slab priors will be adopted to impose exact sparsity on regression coefficients. Otherwise, Laplacian shrinkage will be adopted. The default value is TRUE.
 #' @param model the model to be fitted. Users can specify "linear" for a linear model and "VC" for a varying coefficient model.
-#' @param hyper a named list of hyperparameters. The default value is NULL.
+#' @param hyper a named list of hyper-parameters. The default value is NULL.
 #' @param debugging logical flag. If TRUE, progress will be output to the console and extra information will be returned. The default value is FALSE.
 #'
 #' @details The quantile regression model described in "\code{\link{data}}" is:
 #' \deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}X_{ij}\gamma_j +\epsilon_{i},}
-#' where \eqn{\beta_k}'s are the regression coefficients for the clinical covariates (including the intercept if specified) and \eqn{\gamma_j}'s are the regression coefficients of \eqn{\boldsymbol X}.
+#' where \eqn{\beta_k}'s are the regression coefficients for clinical covariates and \eqn{\gamma_j}'s are the regression coefficients of \eqn{\boldsymbol X}.
 #' 
 #' The quantile varying coefficient model described in "\code{\link{data}}" is:
 #' \deqn{Y_{i}=\sum_{k=1}^{q} E_{ik} \beta_k +\sum_{j=0}^{p}\gamma_j(V_i)X_{ij} +\epsilon_{i},}
-#' where \eqn{\beta_k}'s are the regression coefficients for the clinical covariates and \eqn{\gamma_j}'s are the varying coefficients for the intercept and predictors (e.g. genetic factors).
+#' where \eqn{\beta_k}'s are the regression coefficients for the clinical covariates, and \eqn{\gamma_j}'s are the varying intercept and varying coefficients for predictors (e.g. genetic factors), respectively.
 #'
 #' 
 #' Users can modify the hyper-parameters by providing a named list of hyper-parameters via the argument `hyper'.
@@ -45,17 +45,17 @@
 #' y=data$y
 #' e=data$e
 #' 
-#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,model="linear")
+#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,model="linear")
 
 #' \donttest{
 #'
 #' ## non-sparse
 #' sparse=FALSE
-#' fit2=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,sparse = sparse,model="linear")
+#' fit2=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,sparse = sparse,model="linear")
 #' 
 #' ## non-robust
 #' robust = FALSE
-#' fit3=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,robust = robust,model="linear")
+#' fit3=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,robust = robust,model="linear")
 #' 
 #' }
 #' 
@@ -67,17 +67,18 @@
 #' u=data$u
 #' e=data$e
 #'
-#' fit1=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,model="VC")
+#' spline = list(kn=2,degree=2)
+#' fit1=pqrBayes(g,y,u,e,quant=0.5,spline = spline, model="VC")
 #'
 #' \donttest{
 #'
 #' ## non-sparse
 #' sparse=FALSE
-#' fit2=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,sparse = sparse,model="VC")
+#' fit2=pqrBayes(g,y,u,e,quant=0.5,spline = spline,sparse = sparse,model="VC")
 #' 
 #' ## non-robust
 #' robust = FALSE
-#' fit3=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,robust = robust,model="VC")
+#' fit3=pqrBayes(g,y,u,e,quant=0.5,spline = spline,robust = robust,model="VC")
 #' 
 #' }
 #' @export
@@ -95,12 +96,13 @@
 #' {\emph{Statistics in Medicine}, 39: 617– 638} \doi{10.1002/sim.8434}
 
 
-pqrBayes <- function(g, y, u=NULL, e,quant=0.5, iterations=10000, kn=NULL, degree=NULL, robust=TRUE,sparse=TRUE, model="linear",hyper=NULL,debugging=FALSE){
+pqrBayes <- function(g, y, u=NULL, e,quant=0.5, iterations=10000, burn.in=NULL,spline, robust=TRUE,sparse=TRUE, model="linear",hyper=NULL,debugging=FALSE){
   
   if(model=="VC"){
-    fit = pqrBayes_vc(g, y, u, e,quant, iterations, kn, degree,robust,sparse, hyper,debugging)
+    kn = as.integer(spline$kn); degree = as.integer(spline$degree)
+    fit = pqrBayes_vc(g, y, u, e,quant, iterations, burn.in, kn, degree,robust,sparse, hyper,debugging)
   }else if(model=="linear"){
-    fit = pqrBayes_lin(g, y, e,quant, iterations, robust,sparse, hyper,debugging)
+    fit = pqrBayes_lin(g, y, e,quant, iterations, burn.in, robust,sparse, hyper,debugging)
   }
   else{
     stop("model should be either VC or linear")