+++ R/select.pqrBayes.R
@@ -30,7 +30,7 @@
 #' y=data$y
 #' e=data$e
 #' 
-#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,kn=NULL,degree=NULL,model="linear")
+#' fit1=pqrBayes(g,y,u=NULL,e,quant=0.5,spline=NULL,model="linear")
 #' sparse=TRUE
 #' select=pqrBayes.select(obj = fit1,sparse = sparse,model="linear")
 #' 
@@ -41,7 +41,8 @@
 #' y=data$y
 #' u=data$u
 #' e=data$e
-#' fit1=pqrBayes(g,y,u,e,kn=2,degree=2,quant=0.5,model="VC")
+#' spline = list(kn=2,degree=2)
+#' fit1=pqrBayes(g,y,u,e,quant=0.5,spline = spline, model="VC")
 #' sparse=TRUE
 #' select=pqrBayes.select(obj = fit1,sparse = sparse,model="VC")
 #' select
@@ -49,7 +50,8 @@
 #' \donttest{
 #' ## non-sparse
 #' sparse=FALSE
-#' fit2=pqrBayes(g,y,u,e,quant=0.5,kn=2,degree=2,sparse = sparse,model="VC")
+#' spline = list(kn=2,degree=2)
+#' fit2=pqrBayes(g,y,u,e,quant=0.5,spline = spline,sparse = sparse,model="VC")
 #' select=pqrBayes.select(obj=fit2,sparse=FALSE,model="VC")
 #' select
 #' }