+++ NEWS.md
@@ -1,9 +1,16 @@
 
 ## News
 
+### pqrBayes 1.1.1 \[2025-02-22\]
+
+- Updated pqrBayes function with burn-in and spline parameters.
+- Fixed bugs in the simulation example under high-dimensional linear regression model.
+- Clarified the intercept in sparse linear and VC models in the User's manual.
+
 ### pqrBayes 1.1.0 \[2025-02-16\]
 
-- Added functions to compute empirical coverage probabilities of sparse regression coefficients under linear models. 
+- Added functions to fit robust sparse Bayesian linear models and their non-robust counterparts. 
+- Added functions to compute empirical coverage probabilities of sparse regression coefficients under linear models.
 - Added examples to reproduce the inference results from Fan et al. (2024).
 
 ### pqrBayes 1.0.5 \[2025-02-05\]
@@ -27,6 +34,7 @@
 ### pqrBayes 1.0.2 \[2023-09-14\]
 
 - Initial release to CRAN.
+- Included functions to fit robust sparse VC models (with or without spike-and-slab priors).